backend/requirements.txt
asgiref==3.10.0
attrs==25.4.0
colorama==0.4.6
Django==5.2.8
django-autoslug==1.9.9
django-cities==0.6.2
django-cors-headers==4.9.0
django-countries==8.1.1
django-filter==25.2
django-model-utils==5.0.0
django-phonenumber-field==8.3.0
django-stubs==5.2.7
django-stubs-ext==5.2.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
phonenumbers==9.0.18
pillow==12.0.0
pluggy==1.6.0
progressbar2==4.5.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.1
python-utils==3.9.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.29.0
six==1.17.0
sqlparse==0.5.3
swapper==1.4.0
tqdm==4.67.1
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
tzdata==2025.2
Unidecode==1.4.0
uritemplate==4.2.0
